+++ src/deps.cljs
@@ -1,3 +1,3 @@
 {:npm-deps
- {"@cortex-js/compute-engine" "^0.12.1",
-  "@mentatcollective/mathlive" "0.85.2"}}
+ {"@cortex-js/compute-engine" "^0.12.2",
+  "mathlive" "^0.86.0"}}