+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
-## unreleased
+## [unreleased]
 
-- More soon.
+## 0.1.0
+
+- Added final `Mathfield component`
+
+- Published fleshed-out documentation notebook at https://mentat-collective.github.io/mathlive.cljs
+
+- Upgraded `mathlive` to `0.86.0` and `compute-engine` to `0.12.2`
 
 ## 0.0.1
 
-First release!
+- First release!